requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
cloudinary==1.33.0
colorama==0.4.6
daraja-mpesa==1.2.1
dj-database-url==2.1.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-cron==0.6.0
django-daraja==1.1.2
django-dotenv==1.4.2
django-filter==23.2
django-hashid-field==3.3.7
django-heroku==0.3.1
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
frozenlist==1.4.0
future==0.18.3
hashids==1.3.1
idna==3.4
multidict==6.0.4
packaging==23.2
Pillow==9.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.6.0
pypng==0.20220715.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
twilio==8.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.6.0
yarl==1.9.3
